Review & Comment only || Object overhaul - POC  (#339)
* Give property enums a static value to aviod positional fubar later on.   Cleaned up using.

* Cleaning up enums

* Fixed Line ending update change log.

* Added new positionType

* Started work on implementing object overhaul - want to make sure I am heading in the right direction.

* Minor clean up, responded to feedback and fixed a few bugs.   Thanks for the second and third set of eyes.

+++ Source/ACE/Entity/Creature.cs
@@ -58,8 +58,8 @@ namespace ACE.Entity
         }
 
         public Creature(AceCreatureStaticLocation aceC)
-            : base((ObjectType)aceC.CreatureData.TypeId, 
-                  new ObjectGuid(CommonObjectFactory.DynamicObjectId, GuidType.Creature), 
+            : base((ObjectType)aceC.CreatureData.TypeId,
+                  new ObjectGuid(CommonObjectFactory.DynamicObjectId, GuidType.Creature),
                   aceC.CreatureData.Name,
                   aceC.WeenieClassId,
                   (ObjectDescriptionFlag)aceC.CreatureData.WdescBitField,
@@ -83,7 +83,7 @@ namespace ACE.Entity
             PhysicsData.CSetup = aco.ModelTableId;
             PhysicsData.Petable = aco.PhysicsTableId;
             PhysicsData.ObjScale = aco.ObjectScale;
-            
+
             // this should probably be determined based on the presence of data.
             PhysicsData.PhysicsDescriptionFlag = (PhysicsDescriptionFlag)aco.PhysicsBitField;
             PhysicsData.PhysicsState = (PhysicsState)aco.PhysicsState;
@@ -91,7 +91,8 @@ namespace ACE.Entity
             // game data min required flags;
             Icon = aco.IconId;
 
-            GameData.ContainerCapacity = aco.ContainersCapacity;
+            // TODO: Check to see if we should default a 0 to fix these possible null errors Og II
+            GameData.ContainerCapacity = (byte)aco.ContainersCapacity;
             GameData.ItemCapacity = aco.ItemsCapacity;
             GameData.Usable = (Usable)aco.Usability;
             // intersting finding: the radar color is influenced over the weenieClassId and NOT the blipcolor
@@ -172,7 +173,7 @@ namespace ACE.Entity
             UniversalMotion motionDeath = new UniversalMotion(MotionStance.Standing, new MotionItem(MotionCommand.Dead));
             QueuedGameAction actionDeath = new QueuedGameAction(this.Guid.Full, motionDeath, 2.0f, true, GameActionType.MovementEvent);
             session.Player.AddToActionQueue(actionDeath);
-            
+
             // Create Corspe and set a location on the ground
             // TODO: set text of killer in description and find a better computation for the location, some corpse could end up in the ground
             var corpse = CorpseObjectFactory.CreateCorpse(this, this.Location);
@@ -183,7 +184,7 @@ namespace ACE.Entity
             // corpse.DespawnTime = Math.Max((int)session.Player.PropertiesInt[Enum.Properties.PropertyInt.Level] * 5, 360) + WorldManager.PortalYearTicks; // as in live
             corpse.DespawnTime = 20 + WorldManager.PortalYearTicks; // only for testing
 
-            // If the object is a creature, Remove it from from Landblock 
+            // If the object is a creature, Remove it from from Landblock
             if (!isDerivedPlayer)
             {
                 QueuedGameAction removeCreature = new QueuedGameAction(this.Guid.Full, this, true, true, GameActionType.ObjectDelete);
